[math] Fix matrix_from_function for stacked outputs
* Add default value to Shape.get_size(dim)
* Improve sparse tensor formatting

+++ tests/commit/math/test__sparse.py
@@ -14,9 +14,9 @@ class TestSparse(TestCase):
         self.assertEqual(1, get_sparsity(wrap(1)))
         self.assertEqual(0.25, get_sparsity(expand(1., batch(b=4))))
         self.assertEqual(0.25, get_sparsity(stack([zeros(batch(b=4))] * 3, channel('vector'))))
-        self.assertEqual(0.3, get_sparsity(SparseCoordinateTensor(ones(instance(nnz=3), channel(vector='x')), ones(instance(nnz=3)), spatial(x=10), True, False)))
-        self.assertEqual(0.03, get_sparsity(CompressedSparseMatrix(indices=ones(instance(nnz=3)),
-                                                                   pointers=ones(instance(y_pointers=11)),
+        self.assertEqual(0.3, get_sparsity(SparseCoordinateTensor(ones(instance(nnz=3), channel(vector='x'), dtype=int), ones(instance(nnz=3)), spatial(x=10), True, False)))
+        self.assertEqual(0.03, get_sparsity(CompressedSparseMatrix(indices=ones(instance(nnz=3), dtype=int),
+                                                                   pointers=ones(instance(y_pointers=11), dtype=int),
                                                                    values=ones(instance(nnz=3)),
                                                                    uncompressed_dims=spatial(x=10),
                                                                    compressed_dims=spatial(y=10))))
